audiobook_backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
websockets==13.1
httpx==0.28.1

# ── Data / validation ─────────────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0          # PostgreSQL async driver

# ── Task queue ───────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.2.1
flower==2.0.1             # Celery monitoring UI

# ── Document parsing ──────────────────────────────────────────────────────────
PyMuPDF==1.25.1
python-docx==1.1.2
ebooklib==0.18
beautifulsoup4==4.12.3
lxml==5.3.0

# ── NLP ──────────────────────────────────────────────────────────────────────
spacy==3.8.3

# ── TTS ──────────────────────────────────────────────────────────────────────
elevenlabs==1.14.0

# ── Audio ─────────────────────────────────────────────────────────────────────
pydub==0.25.1

# ── Auth / Security ───────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ── Utilities ────────────────────────────────────────────────────────────────
aiofiles==24.1.0
python-dotenv==1.0.1
